chore (hooks): Rename datable hook file name
